Added frontend and refactored some logic

+++ src/main/java/com/spring/controllers/ProductSiteController.java
@@ -30,7 +30,7 @@ public class ProductSiteController {
     @PostMapping(path = "/add")
     public Product addProduct(@RequestParam String name, @RequestParam String description, @RequestParam String brand,
                            @RequestParam String material, @RequestParam String manufacturerCountry,
-                           @RequestParam String category, @RequestParam int price, @RequestParam float weight) {
+                           @RequestParam String category, @RequestParam int price, @RequestParam double weight) {
         if (!name.isEmpty() && !description.isEmpty() && !material.isEmpty() && !brand.isEmpty() &&
                 !manufacturerCountry.isEmpty() && !category.isEmpty() && price >= 0 && weight > 0)
             return service.addProduct(new Product(name, description, brand, material, manufacturerCountry, category, price, weight));
